add project display view w/back btn
